requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai-whisper==20231117
openai==1.3.5
torch==2.1.1
torchvision==0.16.1
clip-by-openai==1.0
moviepy==1.0.3
opencv-python==4.8.1.78
Pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy>=1.24.3,<2.3.0
